% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{claims}
\alias{claims}
\title{claims}
\description{
In the data set, all the claimants are male, between the ages of 35 and 54 years, work in the logging industry and
reported their claim to the Richmond, BC Workers Compensation Board. Only claimants whose
injuries were due to cuts and lacerations were included in the data set.
}
\references{
Freeland RK. Statistical analysis of discrete time series with applications to the analysis of workers compensation
claims data [unpublished doctoral dissertation]. Vancouver (Canada): University of British Columbia; 1998.
}

