% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inar.R
\name{coef.inar}
\alias{coef.inar}
\title{Function coef.inar}
\usage{
\method{coef}{inar}(object, ...)
}
\arguments{
\item{object}{An object of class "inar".}
}
\value{
Resultados
}
\description{
Description
}
\references{

}

