% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inar.R
\name{fitted.inar}
\alias{fitted.inar}
\title{Function fitted.inar}
\usage{
\method{fitted}{inar}(object, ...)
}
\arguments{
\item{object}{An object of class "inar".}
}
\value{
Resultados
}
\description{
Description
}
\references{

}

