% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{nginar.sim}
\alias{nginar.sim}
\title{Function nginar.sim}
\usage{
nginar.sim(n, alpha, mu, n.start = 150)
}
\arguments{
\item{n}{A strictly positive integer.}

\item{alpha}{a vector of INAR coefficients.}

\item{mu}{a vector.}

\item{n.start}{the length of 'burn-in' period. If na, the default, a reasonable valve is computed.}
}
\value{
A time-series object of class "ts".
}
\description{
Simulate from an Inar model
}
\examples{

# A new geometric INAR simulation
ts.sim <- nginar.sim(n = 100, alpha = 0.4, mu = 2)
ts.plot(ts.sim)

}
\references{
Ristic, M.M., Bakouchb, H.S. and Nasti, A.S. (2009) 
A newgeometricfirst-orderinteger-valuedautoregressive (NGINAR(1)) process . 
\emph{Journal of Statistical Planning and Inference}, \bold{139}, 2218--2226.
}
\seealso{
\code{\link{nginar}}
}

