% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesistests.R
\name{otest}
\alias{otest}
\title{Function Test}
\usage{
otest(x, conf.level = 0.05)
}
\arguments{
\item{x}{a numeric vector or time series.}
}
\value{

}
\description{
Tests
}
\examples{

}
\references{

}

