% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inar.R
\name{plot.inar}
\alias{plot.inar}
\title{Function plot.inar}
\usage{
\method{plot}{inar}(x, ask = interactive(), ...)
}
\arguments{
\item{x}{An object of class "inar".}
}
\value{
Resultados
}
\description{
Description
}
\references{

}

