% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inar.R
\name{poinar}
\alias{poinar}
\title{Function poinar}
\usage{
poinar(x, order.max, series = NULL)
}
\arguments{
\item{x}{a numeric vector or time series.}

\item{order.max}{a one dimensional integer vestor giving the order of the model to fit. This value corresponds the INAR order.}

\item{series}{name for the series.}
}
\value{
Resultados
}
\description{
Fit in inar model to a univariate time series by yule-walker method
}
\examples{
data(claims)
mean(claims[,5])
var(claims[,5])
var(claims[,5])/mean(claims[,5])  # dispersion index
acf(claims[,5])
pacf(claims[,5])
poinar(claims[,5], 1)

}
\references{
Du, J.G. and Li,Y.(1991):
The integer-valued autorregressive (INAR(p)) model.
\emph{Journal of time series analysis} \bold{12}, 129--142.

Freeland RK. Statistical analysis of discrete time series with applications to the analysis of workers compensation
claims data [unpublished doctoral dissertation]. Vancouver (Canada): University of British Columbia; 1998.
}

