% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{poinar.sim}
\alias{poinar.sim}
\title{Function poinar.sim}
\usage{
poinar.sim(n, order.max, alpha, lambda, n.start = NA)
}
\arguments{
\item{n}{the length of outputs series. A strictly positive integer.}

\item{order.max}{the integer component p is the INAR order.}

\item{alpha}{a vector of INAR coefficients.}

\item{lambda}{the mean of the poisson distribution.}

\item{n.start}{the length of 'burn-in' period. If na, the default, a reasonable valve is computed.}
}
\value{
A time-series object of class "ts".
}
\description{
Simulate from an INAR model
}
\examples{
# A Poisson INAR simulation
ts.sim <- poinar.sim(n = 100, order.max = 2, alpha = c(0.1,0.4),lambda = 2, n.start=200)
ts.plot(ts.sim)

}
\references{
Du, J.G. and Li,Y. (1991).
 The integer-valued autorregressive (INAR(p)) model.
 \emph{Journal of time series analysis}. \bold{12}, 129--142.
}
\seealso{
\code{\link{poinar}}
}

