% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inar.R
\name{print.inar}
\alias{print.inar}
\title{Function print.inar}
\usage{
\method{print}{inar}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An object of class "inar".}
}
\value{
Resultados
}
\description{
Description
}
\references{

}

