% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{print.otest}
\alias{print.otest}
\title{Function print.otest}
\usage{
\method{print}{otest}(x, digits = max(3, getOption("digits") - 2),
  signif.stars = getOption("show.signif.stars"), ...)
}
\arguments{
\item{x}{}
}
\value{
Resultados
}
\description{
Description
}
\references{

}

