% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inar.R
\name{residual.inar}
\alias{residual.inar}
\title{Function residuals.inar}
\usage{
residual.inar(object, ...)
}
\arguments{
\item{object}{An object of class "inar".}
}
\value{
Resultados
}
\description{
Description
}
\references{

}

