% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inar.R
\name{residuals.inar}
\alias{residuals.inar}
\title{Function residuals.inar}
\usage{
\method{residuals}{inar}(object, ...)
}
\arguments{
\item{object}{}
}
\value{
Resultados
}
\description{
Description
}
\references{

}

