% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sexoffences}
\alias{sexoffences}
\title{sexoffences}
\description{
The data set is obtained from the crime data section of the forecasting principles
site (http://www.forecastingprinciples.com). This data series represents the counting
o sex offences reported in the 21st police car beat in Pittsburgh, during one
month. The data consist of 144 observations, starting in January 1990 and ending
in December 2001.
}
\references{
Ristic, M. M., Bakouch, H. S. and Nastic, A. S. (2009). A new geometric first-order integer valued
autoregressive (NGINAR(1)) process. Journal of Statistical Planning and Inference 139, 2218 – 2226.
}

