% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{skinlesions}
\alias{skinlesions}
\title{skinlesions}
\description{
Our data give numbers of submissions to animal health laboratories, monthly 2003–2009, from a
region in New Zealand. The submissions can be categorized in various ways. Here we consider one series giving the total
number of bovine cases, and several others categorized by presenting symptoms.
}
\references{
Jazi, M.A, Jones, G. and Lai, C.-D. (2012). First-order integer valued AR processes with
zero in ated Poisson innovations. Journal of Time Series Analysis, 33, 954-963.
}

