% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{soap}
\alias{soap}
\title{soap}
\description{
Consider the series of weekly sales (in integer units) of particular soap product
in a supermarket. The data are taken from a database provided by the Kilts Center
for Marketing, Graduate School of Business of the University of Chicago,
at: http://gbswww.uchicago.edu/kilts/research/db/dominicks. (The product is 'Zest White Water 15 oz.', with code 3700031165)
}
\references{
Iain L. MacDonald and Walter Zucchini (1997). Hidden Markov and Other Models.
Monograph on Statistics and Applied Probability 70.
}

