% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{violence}
\alias{violence}
\title{violence}
\description{
Their original data counts family violence
in the 11th police car beat in Pittsburgh, during one month. The data setis obtained from the crime data section of the forecasting principles site
(http://www.forecastingprinciples.com). It consists of 144 observations, starting in January 1990 and ending in December 2001.
}
\references{
Bakouch, S. H. and Ristic, M. M. (2010). Zero truncated Poisson integer valued AR(1) model.
Metrika 72, 265-280.
}

