% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inar.R
\name{coef.inar}
\alias{coef.inar}
\title{Methods for Fitted INAR Models}
\usage{
## S3 method for class 'inar'
coef(object, ...)
## S3 method for class 'inar'
residuals(object, ...)
## S3 method for class 'inar'
fitted(object, ...)
## S3 method for class 'inar'
print(x, digits = max(3, getOption("digits") - 3), ...)
## S3 method for class 'inar'
plot(x, ask = interactive(), ...)
}
\arguments{
\item{object, }{x an object of class "inar"; usually, a result of a call to inar.}

\item{...}{further arguments passed to or from other methods.}

\item{digits}{see \code{\link{printCoefmat}}.}

\item{ark}{Should the plot method work interactively? See \code{\link{interactive}}.}
}
\value{
For coef, a numeric vector; 
for residuals and fitted a univariate time series; 
for plot and print, the fitted INAR model object.
}
\description{
Methods for fitted INAR model objects.
}

