% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{nginar}
\alias{nginar}
\title{Function nginar}
\usage{
nginar(x, series = NULL)
}
\arguments{
\item{x}{a numeric vector or time series.}

\item{order.max}{the integer component p is the INAR order.}

\item{mu}{the mean of the geometric distribution.}
}
\value{

}
\description{
Fit in inar model to a univariate time series by yule-walker method
}
\examples{

data(sexoffences)
mean(sexoffences)
var(sexoffences)
nginar(sexoffences)

}

