% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{nginar}
\alias{nginar}
\title{Function nginar}
\usage{
nginar(x, series = NULL)
}
\arguments{
\item{x}{a numeric vector or time series.}

\item{order.max}{the integer component p is the NGINAR order.}

\item{mu}{the mean of the geometric distribution.}
}
\value{
A list of class "inar" with the following elements:

\describe{
\item{coef:}{estimated NGINAR coefficients for the fitted model.}
\item{residual:}{the series of residuals.}
\item{fitted.values:}{the fitted series.}
\item{series:}{the name of the series x.}
}
}
\description{
Fit in inar model to a univariate time series by yule-walker method
}
\examples{

data(sexoffences)
mean(sexoffences)
var(sexoffences)
nginar(sexoffences)

}
\references{
Ristic, M.M., Bakouchb, H.S. and Nasti, A.S. (2009) 
A newgeometricfirst-orderinteger-valuedautoregressive (NGINAR(1)) process . 
\emph{Journal of Statistical Planning and Inference}, \bold{139}, 2218--2226.
}

